refactor(hexagon): extract in another file
